Updated project-structure, removed `/extern/` and moved `va-opt.h` into `/lib/`
