+++ .github/workflows/clang-compatibility-check.yml
@@ -29,7 +29,7 @@ jobs:
       - name: Build Tests
         run: |
           cmake . -G "Unix Makefiles" -D CMAKE_C_COMPILER=clang -D CMAKE_CXX_COMPILER=clang++
-          cmake --build . --target runPblTests
+          cmake --build . --target pbl-tests
       - name: Run Tests
         run: |
-          ./tests/runPblTests
+          ./tests/pbl-tests
