+++ .github/workflows/codecov.yml
@@ -14,10 +14,10 @@ jobs:
       - name: Build GTest Tests
         run: |
           cmake .
-          cmake --build . --target runPblTests -DPBL_COVERAGE=true
+          cmake --build . --target pbl-tests -DPBL_COVERAGE=true
       - name: Run C-Coverage
         run: |
-          ./tests/runPblTests
+          ./tests/pbl-tests
       - name: Generate coverage files
         run: |
           gcov -abcfu ./CMakeFiles/pbl.dir/**/*.gcda ./CMakeFiles/pbl.dir/**/*.gcno
