+++ .github/workflows/gcc-compatibility-checks.yml
@@ -29,7 +29,7 @@ jobs:
       - name: Build Tests
         run: |
           cmake .
-          cmake --build . --target runPblTests
+          cmake --build . --target pbl-tests
       - name: Run Tests
         run: |
-          ./tests/runPblTests
+          ./tests/pbl-tests
