+++ examples/CMakeLists.txt
@@ -1,8 +1,7 @@
-# Setting all files to compile
-set(SOURCE_FILES ./simple-io.c)
-
-# Adding the executable example
-add_executable(pbl-examples ${SOURCE_FILES})
+# Adding the executables for the examples
+add_executable(pbl-example-simple-io ./simple-io.c)
+add_executable(pbl-example-simple-string ./simple-string.c)
 
 # Linking the library into the example
-target_link_libraries(pbl-examples PUBLIC pbl)
+target_link_libraries(pbl-example-simple-io PUBLIC pbl)
+target_link_libraries(pbl-example-simple-string PUBLIC pbl)
