+++ include/libpbl/pbl-apply-macro.h
@@ -8,7 +8,7 @@
 #pragma once
 
 // General Required Header Inclusion
-#include "va-opt.h"
+#include <va-opt/va-opt.h>
 
 #ifndef PBL_INCLUDE_APPLY_MACRO_H
 #define PBL_INCLUDE_APPLY_MACRO_H
