+++ include/libpbl/pbl.h
@@ -23,7 +23,7 @@
 #include "./pbl-main.h"
 
 // Included the additional headers
-#include "./additional/va-opt.h"
+#include <va-opt/va-opt.h>
 
 #ifndef PBL_MODULES_LIBRARY_H
 #define PBL_MODULES_LIBRARY_H
