+++ include/libpbl/types/pbl-types.h
@@ -21,7 +21,7 @@
 
 
 // General Required Header Inclusion
-#include "../additional/va-opt.h"
+#include <va-opt/va-opt.h>
 #include "../mem/pbl-mem.h"
 
 #ifndef PBL_MODULES_TYPES_H
